Windows: Use correct default font when desktopSettingsAware is false
The DirectWrite engine had hardcoded a default font instead of checking
the system settings for this.

Usually this will not matter, since the theme will call
QWindowsFontDatabase::systemDefaultFont() to get the default, but when
this code path is disabled by setting
QGuiApplication::desktopSettingsAware to false, the defaultFont()
function is called. With DirectWrite we would always return a Segoe UI
font of size 16 in this case.

The other databases respect the system settings for the default font,
even when desktopSettingsAware is set to false, so this is the expected
behavior. It is also what the GDI font database does, so this change
will appear as a regression for users upgrading.

[ChangeLog][Windows] Fixed a regression where the default font on
Windows would be larger than before if desktopSettingsAware had been set
to false.

Fixes: QTBUG-129832
Change-Id: I959e7a52680678a6ab767b49ee643fade5e7209a
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit b09fb51e7bb576137f92d4697f9f812f3dcb2b7a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>