Windows: Release old references when invalidating DirectWrite db
When the font database is invalidated, we need to release all
references to populated fonts and clear the lists, so that they
can be repopulated from scratch. This was not really measurable
as long as the contents of the database did not change, but it
was a possible source of errors.

Task-number: QTBUG-129849
Change-Id: Ic7ed4abcb13f13332e8edb77e2039f08da9660fd
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 2ea099b67943634a06bce642e13062f84de7e3be)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>