Android: further simplify permission request callback
Since we don't use the list of strings, don't pass it from Java to C++
at all. This saves us making a copy of the strings. And as we need to
create a QList<PermissionResult> from the list of ints, don't receive
the list of int's as a QList<int>, but as a QJniArray<int>, which means
one less copy of the int-array we get from Java.

Amends 46737c0bc0bb86df87f509628797b5cab184760a.

Change-Id: Idae86174cafff90d56affba6e0862fb5c6668c13
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>