Set clear geometry for tst_QWidget::hoverPosition to remove flakiness
In some cases, root.geometry().topLeft() was QPoint(0,0) resulting in
flakiness when testing the hover position of the cursor.

Set a clear geometry to prevent the geometry of root being flaky, this was in some cases giving the wrong information for the cursor position.

The flakiness is a result from touch events on VMs.

Change-Id: Id3d5099e52216dfcd1c4f3b02a1a60cfce299b0f
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>