rhi: Add size param for internal buffer update used by Quick
Because reusing larger buffers for smaller data is very much a thing
in Qt Quick, depending on the scene. It happens rarely for uniform
buffers, for which this feature was originally invented, but when
vertex/index buffers switch over to being Dynamic, this same code
path is hit, and for those, recycling older QRhiBuffers for newer,
smaller data is common.

Change-Id: I41856adf6b8bb5486d05e186d3007f262e7dc987
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 28469b4dfc039601b898a1cdef873d458fbc19eb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>