Blacklist QColorSpace qfloat16-related tests on VxWorks
tst_QColorSpace::imageConversionOverLargerGamut2 fails on VxWorks on
check: `resultImage.pixelColor(255, 0).redF() > 1.0f`, because
`resultImage.pixelColor(255, 0).redF()` returns `1.0f`. This happens
only for qfloat16 checkes, which indicates some problems with VxWorks
handling of this type, but exact reason is unknown.

We need to run COIN tests for VxWorks in a blocking mode, but some
issues which result in failures are still to be addressed. Due to this,
skip some qfloat16-related tests of QColorSpace which fail on VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130076
Change-Id: Ie2f3e28d2573d0c94bba32d4102e8a9b84a68fcf
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>