macOS: Don't stop display-link once started and window is exposed
We were stopping the display-link we use to drive update-requests as soon
as no windows on the screen needed an update, and then starting it again
once a new updateRequest came in.

However there is a cost to starting the display-link, as it starts a
dedicated display-link thread, which resulted in a delay of about one
vsync before getting the corresponding update request delivery.

For Qt Quick this is problematic, because it drives all visual changes
to the scene via updateRequest, so e.g. clicking a MouseArea to change
the color of a Rectangle would not be reflected on screen as fast as it
could.

More problematically, when advancing animations from a timer, which
happens if there is more than one window shown, we would conclude after
polishAndSync that no window needed updates, and stop the display link.
But then the animation timer kicked in and would advance animations,
resulting in an updateRequest, and we'd start the display link up again,
causing janky animations:

 - deliverUpdateRequest
 - polishAndSync, without advancing animations
 - stop display link
 - timer callback, advancing animations
 - requestUpdate
 - start display link
 - timer callback, advancing animations
 - requestUpdate
 - display link callback
 - deliverUpdateRequest

We now keep the display link running once it has been started, and
only stop it if the window is de-exposed (or destroyed) without any
pending update requests.

To avoid overhead of running logic related to update requests when
there's no need, we keep track of whether there is a window on the
screen that needs an updateRequest, and skip the signaling from the
display-link thread over to the main thread if there are none.

We could have moved the bail-out later, to the main thread before
we iterate the windows, or even not tracked pending update requests
at all, which would have given us even faster "response-time", but
doing so would only buy us a small benefit, at the cost of always
signaling the main thread and running a GCD callback. Moving to
the new CADisplayLink API available in macOS 14+ might help solve
this, as the callback always comes in on the same (main) thread.

[ChangeLog][macOS] Display-links are now kept running, even when
windows are not animating, as long as they are visible on screen.
This improves the responsiveness in fulfilling updateRequest with
a corresponding UpdateRequest event, which affects how fast a Qt
Quick window can render the result of a property change.

Task-number: QTBUG-129839
Change-Id: Ieebb60e0f619974849e72be1a41aec66508a1475
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>