CMake: Fix DESTDIR handling in SBOM checksum verification
The DESTDIR env var needs to be evaluated at install time, not
configure time, so the $ENV{} needs to be escaped in the generated
script.

Also improve the error message a bit, to say which path exactly it
expected to read.

This fixes SBOM generation for the debian package builds.

Pick-to: 6.8
Task-number: QTBUG-122899
Change-Id: I6869f0054601dcb35ae7ad7b55f808df4dc12311
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>