Change for QTRY_VERIFY2 in manual test largerThanScreen_QTBUG30142
Auto test tst_QWidget::largerThanScreen_QTBUG30142 is flaky, where the
value of widget.frameGeometry().y() is sometimes -28, when it should be >= 0.

This value is always right before qWaitForWindowExposed() is called.

Replace QVERIFY2 with QTRY_VERIFY2 this gives the XCB thread enough
time to proceed and the test is not flaky anymore.

Change-Id: Idef5485938fb122a780006c51779caef3617ad02
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>