Blacklist QNetworkInterface::localAddress test on VxWorks
`tst_QNetworkInterface::localAddress` test fails on VxWorks on
`'socket.waitForConnected(5000)' returned FALSE. ()` for IP6 addresses.
This suggests some configuration issue on VxWorks, investigation is
ongoing.

We need to run COIN tests for VxWorks in a blocking mode, but some
issues which result in failures are still to be addressed. Due to this,
blacklist QNetworkInterface IP6-related tests  which fail on VxWorks.

Task-number: QTBUG-115777
Task-number: QTBUG-130070
Change-Id: Id1dacb0c441f2bcdf1310f172026dc66d5b9aa57
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>