QImage: fix documentation of toCGImage
Remove the \brief section, which qdoc doesn't support for functions.
Explicitly use \fn to link the documentation, as qdoc doesn't always see
the declaration.

This makes the function not only public, but also documented API,
similar to QImage::toHBITMAP and toHICON. In practice, this is what
was intended, so fixing this in stable branches as well.

Pick-to: 6.5
Change-Id: Id825e494d13f10064742318d2b3bbd8f8f7a1435
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 26375032c92b8729ca5625c225e99a696d7b6b9e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>