Add Qt::ImhDecimaNumbersOnly for decimal numbers in virtual keyboards
iOS maps to UIKeyboardTypeDecimalPad
Andoid maps to TYPE_NUMBER_FLAG_DECIMAL only

This improvement is an important change for iOS, on Android
we can get a numerial pad with a few extra keys.

According to
https://developer.apple.com/design/human-interface-guidelines/virtual-keyboards
the UIKeyboardTypeDecimalPad provides a numerical PAD with the decimal
ponctuation, while currently the best option is:
(Qt::ImhPreferNumbers | Qt::ImhFormattedNumbersOnly)
which will provide UIKeyboardTypeNumbersAndPunctuation but this
iOS keyboard type has all sorts of buttons like ?, !, @, $, @...

So it's harder to type when you just need the decimal ponctuation.

Change-Id: I08d317816eb7a7abd1b7606fcee23e87f30f1936
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>